Add children/parents count to lineage. Refactoring
