+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/controller/DataEntityController.java
@@ -27,9 +27,10 @@ import org.opendatadiscovery.oddplatform.api.contract.model.OwnershipUpdateFormD
 import org.opendatadiscovery.oddplatform.api.contract.model.Tag;
 import org.opendatadiscovery.oddplatform.api.contract.model.TagsFormData;
 import org.opendatadiscovery.oddplatform.api.contract.model.TermRef;
-import org.opendatadiscovery.oddplatform.dto.LineageStreamKind;
+import org.opendatadiscovery.oddplatform.dto.lineage.LineageStreamKind;
 import org.opendatadiscovery.oddplatform.service.AlertService;
 import org.opendatadiscovery.oddplatform.service.DataEntityService;
+import org.opendatadiscovery.oddplatform.service.LineageService;
 import org.opendatadiscovery.oddplatform.service.OwnershipService;
 import org.opendatadiscovery.oddplatform.service.term.TermService;
 import org.springframework.http.ResponseEntity;
@@ -48,15 +49,18 @@ public class DataEntityController
     private final OwnershipService ownershipService;
     private final AlertService alertService;
     private final TermService termService;
+    private final LineageService lineageService;
 
     public DataEntityController(final DataEntityService entityService,
                                 final OwnershipService ownershipService,
                                 final AlertService alertService,
-                                final TermService termService) {
+                                final TermService termService,
+                                final LineageService lineageService) {
         super(entityService);
         this.ownershipService = ownershipService;
         this.alertService = alertService;
         this.termService = termService;
+        this.lineageService = lineageService;
     }
 
     @Override
@@ -234,7 +238,7 @@ public class DataEntityController
     public Mono<ResponseEntity<DataEntityLineage>> getDataEntityDownstreamLineage(final Long dataEntityId,
                                                                                   final Integer lineageDepth,
                                                                                   final ServerWebExchange exchange) {
-        return entityService
+        return lineageService
             .getLineage(dataEntityId, lineageDepth, LineageStreamKind.DOWNSTREAM)
             .subscribeOn(Schedulers.boundedElastic())
             .map(ResponseEntity::ok);
@@ -244,7 +248,7 @@ public class DataEntityController
     public Mono<ResponseEntity<DataEntityLineage>> getDataEntityUpstreamLineage(final Long dataEntityId,
                                                                                 final Integer lineageDepth,
                                                                                 final ServerWebExchange exchange) {
-        return entityService
+        return lineageService
             .getLineage(dataEntityId, lineageDepth, LineageStreamKind.UPSTREAM)
             .subscribeOn(Schedulers.boundedElastic())
             .map(ResponseEntity::ok);
@@ -253,7 +257,7 @@ public class DataEntityController
     @Override
     public Mono<ResponseEntity<DataEntityGroupLineageList>> getDataEntityGroupsLineage(final Long dataEntityGroupId,
                                                                                        final ServerWebExchange exch) {
-        return entityService
+        return lineageService
             .getDataEntityGroupLineage(dataEntityGroupId)
             .subscribeOn(Schedulers.boundedElastic())
             .map(ResponseEntity::ok);
