--- odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/LineageStreamKind.java
@@ -1,7 +1,0 @@
-package org.opendatadiscovery.oddplatform.dto;
-
-public enum LineageStreamKind {
-    UPSTREAM,
-    DOWNSTREAM,
-    FULL_GRAPH
-}
