+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/alert/ExternalAlert.java
@@ -1,4 +1,4 @@
-package org.opendatadiscovery.oddplatform.dto;
+package org.opendatadiscovery.oddplatform.dto.alert;
 
 import java.net.URI;
 import java.time.LocalDateTime;
