+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/lineage/LineageDepth.java
@@ -1,4 +1,4 @@
-package org.opendatadiscovery.oddplatform.dto;
+package org.opendatadiscovery.oddplatform.dto.lineage;
 
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
