+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/lineage/LineageNodeDto.java
@@ -1,0 +1,6 @@
+package org.opendatadiscovery.oddplatform.dto.lineage;
+
+import org.opendatadiscovery.oddplatform.dto.DataEntityDimensionsDto;
+
+public record LineageNodeDto(DataEntityDimensionsDto entity, Integer childrenCount, Integer parentsCount) {
+}
