+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/lineage/LineageStreamKind.java
@@ -1,0 +1,6 @@
+package org.opendatadiscovery.oddplatform.dto.lineage;
+
+public enum LineageStreamKind {
+    UPSTREAM,
+    DOWNSTREAM
+}
