+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/dto/metadata/MetadataBinding.java
@@ -1,4 +1,4 @@
-package org.opendatadiscovery.oddplatform.dto;
+package org.opendatadiscovery.oddplatform.dto.metadata;
 
 public record MetadataBinding(long dataEntityId, long metadataFieldId) {
 }
