+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/mapper/MetadataFieldValueMapper.java
@@ -3,7 +3,7 @@ package org.opendatadiscovery.oddplatform.mapper;
 import java.util.Collection;
 import java.util.List;
 import org.opendatadiscovery.oddplatform.api.contract.model.MetadataFieldValue;
-import org.opendatadiscovery.oddplatform.dto.MetadataDto;
+import org.opendatadiscovery.oddplatform.dto.metadata.MetadataDto;
 
 public interface MetadataFieldValueMapper {
     MetadataFieldValue mapDto(final MetadataDto dto);
