+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/repository/GroupEntityRelationRepository.java
@@ -1,8 +1,14 @@
 package org.opendatadiscovery.oddplatform.repository;
 
 import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import org.opendatadiscovery.oddplatform.model.tables.pojos.GroupEntityRelationsPojo;
 
 public interface GroupEntityRelationRepository {
     void createOrUpdateRelations(final Collection<GroupEntityRelationsPojo> pojos);
+
+    Map<String, List<String>> fetchGroupRelations(final Collection<String> childOddrns);
+
+    List<String> getDEGEntitiesOddrns(final long dataEntityGroupId);
 }
