+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/repository/MetadataFieldValueRepository.java
@@ -1,8 +1,7 @@
 package org.opendatadiscovery.oddplatform.repository;
 
 import java.util.List;
-import java.util.Map;
-import org.opendatadiscovery.oddplatform.dto.MetadataDto;
+import org.opendatadiscovery.oddplatform.dto.metadata.MetadataDto;
 import org.opendatadiscovery.oddplatform.model.tables.pojos.MetadataFieldValuePojo;
 
 public interface MetadataFieldValueRepository extends CRUDRepository<MetadataFieldValuePojo> {
