+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/repository/MetadataFieldValueRepositoryImpl.java
@@ -3,12 +3,11 @@ package org.opendatadiscovery.oddplatform.repository;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import org.jooq.DSLContext;
 import org.jooq.InsertSetStep;
 import org.jooq.Record;
-import org.opendatadiscovery.oddplatform.dto.MetadataDto;
+import org.opendatadiscovery.oddplatform.dto.metadata.MetadataDto;
 import org.opendatadiscovery.oddplatform.model.tables.pojos.MetadataFieldPojo;
 import org.opendatadiscovery.oddplatform.model.tables.pojos.MetadataFieldValuePojo;
 import org.opendatadiscovery.oddplatform.model.tables.records.MetadataFieldValueRecord;
