+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/service/DataEntityServiceImpl.java
@@ -16,7 +16,6 @@ import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityClassAndTy
 import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityDetails;
 import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityGroupFormData;
 import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityGroupLineageList;
-import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityLineage;
 import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityList;
 import org.opendatadiscovery.oddplatform.api.contract.model.DataEntityRef;
 import org.opendatadiscovery.oddplatform.api.contract.model.InternalDescription;
@@ -33,8 +32,8 @@ import org.opendatadiscovery.oddplatform.auth.AuthIdentityProvider;
 import org.opendatadiscovery.oddplatform.dto.DataEntityClassDto;
 import org.opendatadiscovery.oddplatform.dto.DataEntityDetailsDto;
 import org.opendatadiscovery.oddplatform.dto.DataEntityDimensionsDto;
-import org.opendatadiscovery.oddplatform.dto.LineageStreamKind;
-import org.opendatadiscovery.oddplatform.dto.MetadataDto;
+import org.opendatadiscovery.oddplatform.dto.lineage.LineageStreamKind;
+import org.opendatadiscovery.oddplatform.dto.metadata.MetadataDto;
 import org.opendatadiscovery.oddplatform.dto.metadata.MetadataKey;
 import org.opendatadiscovery.oddplatform.exception.NotFoundException;
 import org.opendatadiscovery.oddplatform.mapper.DataEntityMapper;
@@ -84,7 +83,7 @@ public class DataEntityServiceImpl
     private final MetadataFieldValueMapper metadataFieldValueMapper;
     private final TagMapper tagMapper;
     private final ReactiveSearchEntrypointRepository reactiveSearchEntrypointRepository;
-    private final ReactiveGroupEntityRelationRepository groupEntityRelationRepository;
+    private final ReactiveGroupEntityRelationRepository reactiveGroupEntityRelationRepository;
     private final ReactiveTermRepository reactiveTermRepository;
     private final ReactiveOwnershipRepository ownershipRepository;
 
@@ -101,7 +100,7 @@ public class DataEntityServiceImpl
                                  final TagMapper tagMapper,
                                  final NamespaceService namespaceService,
                                  final ReactiveDataEntityRepository reactiveDataEntityRepository,
-                                 final ReactiveGroupEntityRelationRepository groupEntityRelationRepository,
+                                 final ReactiveGroupEntityRelationRepository reactiveGroupEntityRelationRepository,
                                  final ReactiveTermRepository reactiveTermRepository,
                                  final ReactiveOwnershipRepository ownershipRepository) {
         super(entityMapper, entityRepository);
@@ -117,9 +116,9 @@ public class DataEntityServiceImpl
         this.tagMapper = tagMapper;
         this.namespaceService = namespaceService;
         this.reactiveDataEntityRepository = reactiveDataEntityRepository;
-        this.groupEntityRelationRepository = groupEntityRelationRepository;
         this.reactiveTermRepository = reactiveTermRepository;
         this.ownershipRepository = ownershipRepository;
+        this.reactiveGroupEntityRelationRepository = reactiveGroupEntityRelationRepository;
     }
 
     @Override
@@ -332,33 +331,12 @@ public class DataEntityServiceImpl
     }
 
     @Override
-    public Mono<DataEntityLineage> getLineage(final long dataEntityId,
-                                              final int lineageDepth,
-                                              final LineageStreamKind lineageStreamKind) {
-        return Mono
-            .fromCallable(() -> entityRepository.getLineage(dataEntityId, lineageDepth, lineageStreamKind))
-            .flatMap(optional -> optional.isEmpty()
-                ? Mono.error(new NotFoundException())
-                : Mono.just(optional.get()))
-            .map(entityMapper::mapLineageDto);
-    }
-
-    @Override
     public Mono<DataEntityList> getDataEntityGroupsChildren(final Long dataEntityGroupId,
                                                             final Integer page,
                                                             final Integer size) {
         return Mono.fromCallable(() -> entityRepository
                 .getDataEntityGroupsChildren(dataEntityGroupId, page, size))
             .map(entityMapper::mapPojos);
-    }
-
-    @Override
-    public Mono<DataEntityGroupLineageList> getDataEntityGroupLineage(final Long dataEntityGroupId) {
-        return Mono.fromCallable(() -> entityRepository.getDataEntityGroupLineage(dataEntityGroupId))
-            .flatMap(optional -> optional.isEmpty()
-                ? Mono.error(new NotFoundException())
-                : Mono.just(optional.get()))
-            .map(entityMapper::mapGroupLineageDto);
     }
 
     private Mono<DataEntityRef> createDEG(final DataEntityGroupFormData formData,
@@ -375,7 +353,7 @@ public class DataEntityServiceImpl
             .flatMap(pojo -> {
                 final List<String> entityOddrns =
                     formData.getEntities().stream().map(DataEntityRef::getOddrn).toList();
-                return groupEntityRelationRepository.createRelations(pojo.getOddrn(), entityOddrns)
+                return reactiveGroupEntityRelationRepository.createRelations(pojo.getOddrn(), entityOddrns)
                     .ignoreElements().thenReturn(pojo);
             })
             .flatMap(this::updateSearchVectors)
@@ -390,9 +368,9 @@ public class DataEntityServiceImpl
         return Mono.just(formData)
             .map(fd -> entityMapper.applyToPojo(fd, namespace, pojo))
             .flatMap(reactiveDataEntityRepository::update)
-            .flatMap(degPojo -> groupEntityRelationRepository.deleteRelationsExcept(degPojo.getOddrn(), entityOddrns)
-                .ignoreElements().thenReturn(degPojo))
-            .flatMap(degPojo -> groupEntityRelationRepository.createRelations(degPojo.getOddrn(), entityOddrns)
+            .flatMap(degPojo -> reactiveGroupEntityRelationRepository
+                .deleteRelationsExcept(degPojo.getOddrn(), entityOddrns).ignoreElements().thenReturn(degPojo))
+            .flatMap(degPojo -> reactiveGroupEntityRelationRepository.createRelations(degPojo.getOddrn(), entityOddrns)
                 .ignoreElements().thenReturn(degPojo))
             .flatMap(this::updateSearchVectors)
             .map(entityMapper::mapRef);
@@ -401,7 +379,7 @@ public class DataEntityServiceImpl
     private Mono<DataEntityPojo> deleteDEG(final DataEntityPojo pojo) {
         return Flux.zip(
             reactiveTermRepository.deleteRelationsWithTerms(pojo.getId()),
-            groupEntityRelationRepository.deleteRelationsForDEG(pojo.getOddrn()),
+            reactiveGroupEntityRelationRepository.deleteRelationsForDEG(pojo.getOddrn()),
             tagService.deleteRelationsForDataEntity(pojo.getId()),
             ownershipRepository.deleteByDataEntityId(pojo.getId())
         ).then(reactiveDataEntityRepository.delete(pojo.getId()));
@@ -410,8 +388,8 @@ public class DataEntityServiceImpl
     private String generateOddrn(final DataEntityPojo pojo) {
         try {
             return oddrnGenerator.generate(ODDPlatformDataEntityGroupPath.builder()
-                    .id(pojo.getId())
-                    .build(), "id");
+                .id(pojo.getId())
+                .build(), "id");
         } catch (Exception e) {
             log.error("Error while generating oddrn for data entity {}", pojo.getId(), e);
             throw new RuntimeException(e);
