+++ odd-platform-api/src/main/java/org/opendatadiscovery/oddplatform/service/metadata/MetadataIngestionServiceImpl.java
@@ -8,8 +8,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.MapUtils;
 import org.opendatadiscovery.oddplatform.annotation.BlockingTransactional;
-import org.opendatadiscovery.oddplatform.dto.MetadataBinding;
 import org.opendatadiscovery.oddplatform.dto.ingestion.IngestionDataStructure;
+import org.opendatadiscovery.oddplatform.dto.metadata.MetadataBinding;
 import org.opendatadiscovery.oddplatform.dto.metadata.MetadataKey;
 import org.opendatadiscovery.oddplatform.dto.metadata.MetadataOrigin;
 import org.opendatadiscovery.oddplatform.dto.metadata.MetadataTypeEnum;
